tabs integracion
