+++ app/renderer/components/barraHerramientas/index.js
@@ -1,0 +1,1 @@
+const eventos = require('./eventos');
